Fixed a bug when having same names
Fixed a bug when both resemblegame and multichoicegame had the same
name, which had to be a number. Instead of trying to parse an integer
and then catching the exception to direct the logic to the correct code
segments, the forms are now passing a gametype parameter, which is then
used in place of the exception handling.